+++ src/main/java/toyproject/personal/englishconversation/controller/ChatController.java
@@ -7,7 +7,7 @@ import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
-import toyproject.personal.englishconversation.controller.dto.ChatGPTRequestDto;
+import toyproject.personal.englishconversation.controller.dto.chatgpt.ChatGPTRequestDto;
 import toyproject.personal.englishconversation.domain.message.GPTMessage;
 import toyproject.personal.englishconversation.service.ChatService;
