+++ src/main/java/toyproject/personal/englishconversation/service/ChatGPTService.java
@@ -10,10 +10,10 @@ import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 import org.springframework.web.client.RestTemplate;
-import toyproject.personal.englishconversation.config.ChatGPTConfig;
-import toyproject.personal.englishconversation.config.ChatGPTContent;
-import toyproject.personal.englishconversation.controller.dto.ChatGPTRequestDto;
-import toyproject.personal.englishconversation.controller.dto.ChatGPTResponseDto;
+import toyproject.personal.englishconversation.config.chatgpt.ChatGPTConfig;
+import toyproject.personal.englishconversation.config.chatgpt.ChatGPTContent;
+import toyproject.personal.englishconversation.controller.dto.chatgpt.ChatGPTRequestDto;
+import toyproject.personal.englishconversation.controller.dto.chatgpt.ChatGPTResponseDto;
 import toyproject.personal.englishconversation.domain.message.GPTMessage;
 
 import java.io.IOException;
