+++ src/main/java/toyproject/personal/englishconversation/service/ChatService.java
@@ -3,7 +3,7 @@ package toyproject.personal.englishconversation.service;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
-import toyproject.personal.englishconversation.controller.dto.ChatGPTRequestDto;
+import toyproject.personal.englishconversation.controller.dto.chatgpt.ChatGPTRequestDto;
 import toyproject.personal.englishconversation.domain.message.GPTMessage;
 
 import java.io.IOException;
