+++ src/test/java/toyproject/personal/englishconversation/exception/ChatGPTExceptionHandlerTest.java
@@ -12,8 +12,9 @@ import org.springframework.boot.test.mock.mockito.MockBean;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
 import org.springframework.test.web.servlet.MockMvc;
+import org.springframework.test.web.servlet.ResultActions;
 import org.springframework.web.client.HttpClientErrorException;
-import toyproject.personal.englishconversation.controller.dto.ChatGPTRequestDto;
+import toyproject.personal.englishconversation.controller.dto.chatgpt.ChatGPTRequestDto;
 import toyproject.personal.englishconversation.service.ChatService;
 
 import java.io.IOException;
@@ -45,10 +46,13 @@ public class ChatGPTExceptionHandlerTest {
         Mockito.when(chatService.processChatRequest(any(ChatGPTRequestDto.class)))
                 .thenThrow(exception);
 
-        // WhenThen
-        mockMvc.perform(post("/chat")
-                        .contentType(MediaType.APPLICATION_JSON)
-                        .content("{}"))
+        // When
+        ResultActions resultActions = mockMvc.perform(post("/chat")
+                .contentType(MediaType.APPLICATION_JSON)
+                .content("{}"));
+
+        // Then
+        resultActions
                 .andExpect(status().is(status.value()))
                 .andExpect(jsonPath("$.code").value(expectedCode))
                 .andExpect(jsonPath("$.message").value(expectedMessage));
@@ -70,10 +74,13 @@ public class ChatGPTExceptionHandlerTest {
         given(chatService.processChatRequest(any(ChatGPTRequestDto.class)))
                 .willThrow(new IOException());
 
-        // WhenThen
-        mockMvc.perform(post("/chat")
-                        .contentType(MediaType.APPLICATION_JSON)
-                        .content("")) // 잘못된 Json 형식 -> Json 매핑 실패
+        // When
+        ResultActions resultActions =  mockMvc.perform(post("/chat")
+                .contentType(MediaType.APPLICATION_JSON)
+                .content("")); // 잘못된 Json 형식 -> Json 매핑 실패
+
+        // Then
+        resultActions
                 .andExpect(status().isInternalServerError())
                 .andExpect(jsonPath("$.code").value("EX"))
                 .andExpect(jsonPath("$.message").value("서버 오류"));
