+++ src/test/java/toyproject/personal/englishconversation/service/ChatServiceTest.java
@@ -4,9 +4,9 @@ import org.junit.jupiter.api.Test;
 import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.springframework.boot.test.context.SpringBootTest;
-import toyproject.personal.englishconversation.config.ChatGPTContent;
-import toyproject.personal.englishconversation.controller.dto.ChatGPTRequestDto;
-import toyproject.personal.englishconversation.controller.dto.Message;
+import toyproject.personal.englishconversation.config.chatgpt.ChatGPTContent;
+import toyproject.personal.englishconversation.controller.dto.chatgpt.ChatGPTRequestDto;
+import toyproject.personal.englishconversation.controller.dto.chatgpt.Message;
 import toyproject.personal.englishconversation.domain.message.GPTMessage;
 
 import java.io.IOException;
